Refactor createHost2Widget schematic and utility functions for improved widget generation
